src/tools/pyproject.toml
[tool.poetry]
name = "tools"
version = "0.3.13"
description = ""
authors = ["mszawerd <mikolaj.szawerda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
sphinx = "^7.3.7"
coverage = "^7.5.0"
flake8 = "^7.0.0"
python-dotenv = "^1.0.1"
webdavclient3 = "^3.14.6"
pyzotero = "^1.5.18"
jupyter = "^1.0.0"
jupyterlab = "^4.1.8"
rich = "^13.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


